fix setComponentValue to transfer long values. the number component accept 32bit values. the int in arduino is 16 bit only

+++ Nextion.h
@@ -70,7 +70,7 @@ class Nextion{
 
   uint8_t buttonOnOff(String find_component, String unknown_component, uint8_t pin, int btn_prev_state);
 
-  boolean setComponentValue(String component, int value);
+  boolean setComponentValue(String component, long value);
 
   //boolean ack(void);//Deprecated
